// src/data/models/Weather.ts
export interface Weather {
  temperature: number;
  humidity: number;
  description: string;
  windSpeed: number;
  weatherCondition: string;
  feelsLike: number;
  city: string;
  country: string;
  sunrise: number;
  sunset: number;
  timezone: number;
  id: number;
  lat: number;
  lon: number;
  icon: string;
}